[Renfi++] Argument handling and validation
In this change we are now fetching the specific
arguments that are used during execution.

We are then validating if the flags are allowed
to be used and terminate the execution if that
is not the case.

Finally, we are also validating the path to make
sure it exists and terminates the execution if
that directory is not available.